Updated main.py
Patch : Reversed keys and values of the dictionnary of registered users, it is to have each IPs associates to name and not the reverse. For which, when receiving a text from an known IP (with an associated name) the text will now prompt : "[From (Name)] : "Their Text" ", which is quite nice to know when you want to discuss with multiple people and know who's talking to you!