Reverted code to earlier version.
Changed the code without test, the next I come back to it and everything about receiving is broken. Since I don't have the time to clear that issue, I had to make the sad choice to revert back to a version that worked. I did some change that before prolly cause myself to do some silly mistakes. But, today's the big day. And I really, like really don't want to mess it up.